requirements.txt
keras==2.13.1
numpy==1.22.3
Pillow==10.0.1
streamlit==1.17.0
streamlit_webrtc==0.47.0
